Refactor: Move story model code to be consistent
